taxhelp-bot/requirements.txt
python-telegram-bot==20.7
pytesseract>=0.3.13
Pillow
opencv-python==4.8.1.78

pandas
numpy>=1.26.0
reportlab==4.0.7
langdetect==1.0.9
googletrans
boto3==1.28.62
cryptography==41.0.5
PyPDF2==3.0.1
python-dotenv==1.0.0
